feat: added user information to avatar menu
