+++ ethereal-nexus-dashboard/src/auth/actions/signInAction.ts
@@ -1,0 +1,7 @@
+'use server'
+
+import { signIn } from '@/auth';
+
+export async function signInAction() {
+  return signIn('Credentials')
+}
