+++ ethereal-nexus-dashboard/src/auth/actions/signOutAction.ts
@@ -1,0 +1,7 @@
+'use server'
+
+import { signOut } from '@/auth';
+
+export async function signOutAction() {
+  return signOut()
+}
